judge adjust yml and judge method

+++ db/migrate/20191218061130_create_quizcards.rb
@@ -8,7 +8,7 @@ class CreateQuizcards < ActiveRecord::Migration[5.1]
       t.string :connotation
       t.string :pronunciation
       t.string :origin
-      t.decimal :wait_seconds
+      t.integer :wait_seconds
       t.references :user, foreign_key: true
       t.timestamps
     end
